Add authentication when chagning or deleting reviews in views.py

+++ presentation/views.py
@@ -1,10 +1,9 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.views import generic
+from django.contrib import messages
 from .models import Item
 from .forms import ItemForm
 from django.core.paginator import Paginator
-# from .forms import LeaveReview, Post
-
 
 
 # Index page
@@ -45,7 +44,7 @@ def leave_review(request):
             
             if form.is_valid():
                 obj = form.save(commit=False) 
-                obj.author_id = request.user.id
+                obj.author = request.user
                 obj.save() 
             else:
                 print("ERROR : Form is invalid")
@@ -62,8 +61,12 @@ def edit_review(request, item_id):
     item = get_object_or_404(Item, id=item_id)
     if request.method == 'POST':
         form = ItemForm(request.POST, instance = item)
-        if form.is_valid():
+        if form.is_valid() and obj.author == request.user:
             form.save()
+        else:
+            messages.add_message(request, messages.ERROR,
+                                 'You can only delete your own review!')
+                                
             return redirect('reviews')
     form = ItemForm(instance=item)
     context = {
@@ -73,8 +76,13 @@ def edit_review(request, item_id):
 
 
 def delete_review(request, item_id):
+    
     item = get_object_or_404(Item, id=item_id)
-    item.delete()
+    if obj.author == request.user:
+        item.delete()
+    else:
+        messages.add_message(request, messages.ERROR,
+                             'You can only delete your own reviews!')
     return redirect('reviews')
